requirements.txt
fastapi==0.135.3
uvicorn[standard]>=0.35.0,<1.0.0
pydantic>=2.12,<3.0
pydantic-settings==2.13.1
python-dotenv>=1.1,<2.0
numpy==2.4.4
faiss-cpu==1.13.2
torch==2.11.0
sentence-transformers==5.3.0
transformers==5.5.0
rank-bm25==0.2.2
Pillow>=11.3,<12.0
PyMuPDF==1.27.2.2
python-multipart==0.0.24
openai==2.30.0
tenacity==9.1.4
pytest>=8.4,<9.0
